+++ c3IDE/Utilities/Extentions/StringExtentions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -33,6 +34,16 @@ namespace c3IDE.Utilities.Extentions
                 @"(\p{Ll})(\P{Ll})",
                 $"$1{rep}$2"
             );
+        }
+
+        public static Stream ToStream (this string s)
+        {
+            var stream = new MemoryStream();
+            var writer = new StreamWriter(stream);
+            writer.Write(s);
+            writer.Flush();
+            stream.Position = 0;
+            return stream;
         }
     }
 }